requirements.txt
streamlit==1.31.1
python-dotenv==1.0.1
requests==2.31.0
PyPDF2==3.0.1
pandas==2.2.1
numpy==1.26.4
google-generativeai==0.3.2 